// frontend/src/data/sdk/backup/backup.gen.ts
export type backupGetAllResponse200TextPlain = {
  data: BackupDTO[];
  status: 200;
};

export type backupGetAllResponse200ApplicationJson = {
  data: BackupDTO[];
  status: 200;
};

export type backupGetAllResponse200TextJson = {
  data: BackupDTO[];
  status: 200;
};

export type backupGetAllResponseSuccess = (
  | backupGetAllResponse200TextPlain
  | backupGetAllResponse200ApplicationJson
  | backupGetAllResponse200TextJson
) & {
  headers: Headers;
};