// frontend/src/data/sdk/dex/dex.gen.ts
export type dexGetAllResponse200TextPlain = {
  data: DexGetAll200One;
  status: 200;
};

export type dexGetAllResponse200ApplicationJson = {
  data: DexGetAll200Two;
  status: 200;
};

export type dexGetAllResponse200TextJson = {
  data: DexGetAll200Three;
  status: 200;
};

export type dexGetAllResponseSuccess = (
  | dexGetAllResponse200TextPlain
  | dexGetAllResponse200ApplicationJson
  | dexGetAllResponse200TextJson
) & {
  headers: Headers;
};