// frontend/src/data/sdk/save-infos/save-infos.gen.ts
export type saveInfosGetAllResponse200TextPlain = {
  data: SaveInfosGetAll200One;
  status: 200;
};

export type saveInfosGetAllResponse200ApplicationJson = {
  data: SaveInfosGetAll200Two;
  status: 200;
};

export type saveInfosGetAllResponse200TextJson = {
  data: SaveInfosGetAll200Three;
  status: 200;
};

export type saveInfosGetAllResponseSuccess = (
  | saveInfosGetAllResponse200TextPlain
  | saveInfosGetAllResponse200ApplicationJson
  | saveInfosGetAllResponse200TextJson
) & {
  headers: Headers;
};