// frontend/src/data/sdk/settings/settings.gen.ts
export type settingsGetResponse200TextPlain = {
  data: SettingsDTO;
  status: 200;
};

export type settingsGetResponse200ApplicationJson = {
  data: SettingsDTO;
  status: 200;
};

export type settingsGetResponse200TextJson = {
  data: SettingsDTO;
  status: 200;
};

export type settingsGetResponseSuccess = (
  | settingsGetResponse200TextPlain
  | settingsGetResponse200ApplicationJson
  | settingsGetResponse200TextJson
) & {
  headers: Headers;
};