// frontend/src/data/sdk/static-data/static-data.gen.ts
export type staticDataGetResponse200TextPlain = {
  data: StaticDataDTO;
  status: 200;
};

export type staticDataGetResponse200ApplicationJson = {
  data: StaticDataDTO;
  status: 200;
};

export type staticDataGetResponse200TextJson = {
  data: StaticDataDTO;
  status: 200;
};

export type staticDataGetResponseSuccess = (
  | staticDataGetResponse200TextPlain
  | staticDataGetResponse200ApplicationJson
  | staticDataGetResponse200TextJson
) & {
  headers: Headers;
};