// frontend/src/data/sdk/storage/storage.gen.ts
export type storageGetBoxesResponse200TextPlain = {
  data: BoxDTO[];
  status: 200;
};

export type storageGetBoxesResponse200ApplicationJson = {
  data: BoxDTO[];
  status: 200;
};

export type storageGetBoxesResponse200TextJson = {
  data: BoxDTO[];
  status: 200;
};

export type storageGetBoxesResponseSuccess = (
  | storageGetBoxesResponse200TextPlain
  | storageGetBoxesResponse200ApplicationJson
  | storageGetBoxesResponse200TextJson
) & {
  headers: Headers;
};
// ...
export type storageGetActionsResponse200TextPlain = {
  data: DataActionPayload[];
  status: 200;
};

export type storageGetActionsResponse200ApplicationJson = {
  data: DataActionPayload[];
  status: 200;
};

export type storageGetActionsResponse200TextJson = {
  data: DataActionPayload[];
  status: 200;
};

export type storageGetActionsResponseSuccess = (
  | storageGetActionsResponse200TextPlain
  | storageGetActionsResponse200ApplicationJson
  | storageGetActionsResponse200TextJson
) & {
  headers: Headers;
};