// frontend/src/data/sdk/warnings/warnings.gen.ts
export type warningsGetWarningsResponse200TextPlain = {
  data: WarningsDTO;
  status: 200;
};

export type warningsGetWarningsResponse200ApplicationJson = {
  data: WarningsDTO;
  status: 200;
};

export type warningsGetWarningsResponse200TextJson = {
  data: WarningsDTO;
  status: 200;
};

export type warningsGetWarningsResponseSuccess = (
  | warningsGetWarningsResponse200TextPlain
  | warningsGetWarningsResponse200ApplicationJson
  | warningsGetWarningsResponse200TextJson
) & {
  headers: Headers;
};